F705 - Lab_Sim/Trabajos/Proyecto02/requirements.txt
pygame
matplotlib
numpy
pygame
pyyaml
mypy
types-PyYAML
